payment api is done

+++ easiness-core/src/main/java/com/easiness/core/repositories/AccountRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import java.util.List;
 import java.util.Optional;
 
-public interface AccountRepository extends JpaRepository<AccountEntity, Integer>, JpaSpecificationExecutor<AccountEntity> {
+public interface AccountRepository extends JpaRepository<AccountEntity, Long>, JpaSpecificationExecutor<AccountEntity> {
     Optional<AccountEntity> findByAccountNo(String accountNo);
     Optional<AccountEntity> findById(Long id);
     Optional<AccountEntity> findByAccountName(String accountName);
